shorten content-id and product id, define uniq fields
When trying to add 'sync' support, I realize dthat there was no easy way to
decide if you'd seen something before.

Now, define a unique tuple for locating content.
 (content-id, product, version)

Those define a unique entity.  This will suffice to determine if new
content is necessary, but there is still no simple way to fully identify
an 'item', in the abscence of a item-unique 'name'.

bzr-revno: 134.1.9