add 'item_id'.
Rather than choosing a static identifier for the unique 'item', the content
can now declare it in 'item_id'.

So, now a unique identifier for an item is:
(content-id, productname, version, item_id).

This isn't actually implemented perfectly for aws, but we easily add a id
ourselves rather than re-using the ami-id.  The problem with using that is that
the same item id could occur in different regions and thus could (unlikely)
collide inside of our version.

bzr-revno: 134.1.10