contentsource: in url content source, late bind by default
Now, UlrContentSource("http://some.url") will not actually do the open.
It adds a 'open' to explicitly do the open, but implicitly does it
inside a read.  This is a bit strange in that 'read' can now raise ENOENT.

but that can be avoided by explicitly calling 'open'.
Also, context calling will force open early.

bzr-revno: 134.2.4