keep certain items from copying up in a condense
Because we are only uploading a single image, its id, name, md5, were seen
to be copy-upable.

Meaning, instead of this:

   "versions": {
    "20121026.1": {
     "items": {
      "disk1.img": {
       "name": "foovendor-pinky-6.1-amd64-server-20121026.1-disk1.img",
       "ftype": "disk1.img",
       "id": "737ed9bc-e111-44eb-8f5e-b55eba4e7095",
       "md5": "da499b246265a34db8576983371c6c2a"
      }

we'd see 
    "20121026.1": {
     "name": "foovendor-pinky-6.1-amd64-server-20121026.1-disk1.img",
     "ftype": "disk1.img",
     "id": "737ed9bc-e111-44eb-8f5e-b55eba4e7095",
     "md5": "da499b246265a34db8576983371c6c2a"
     "items": {
      "disk1.img": { }
     }

Which just looks odd, and its reasonable to expect certain things to *not*
copy up.

Not convinced this is the best solution, but it is one solution.

bzr-revno: 197.1.5