support writing to multiple regions
This supports writing to all regions listed in keystone.
That is useful, but only to some extent.  The reason its less
that useful is that we dont have any cache for the downloading.

Meaning after we do region-a, we would re-download things for region-b.

This could be managed by using 'do-mirror' to maintain a local mirror
and then populating glance from that.  But at some point, ideally
this would somehow cache the downloaded objects locally.

bzr-revno: 215.1.2