modify glance mirror to require content_id
This allows us to create an aggregate list from multiple upstream download
content streams.  As an example:
 sstream-mirror-glance \
   --cloud-name=localcloud "--content-id=localcloud.%(region)s:partners" \
   --output-dir=my.out cirros-mirror/ streams/v1/index.js

Then:
 sstream-mirror-glance \
   --cloud-name=localcloud "--content-id=localcloud.%(region)s:partners" \
   --output-dir=my.out examples/foocloud streams/v1/index.js

That creates an aggregate content_id 'localcloud.Region1:partners'
that contains entries for both cirros and foocloud.

the mirror does not remove products not listed in the source (if it did
then populating 'cirros' would drop 'foocloud' and vice versa.

bzr-revno: 227