glance-mirror: fix issue where 'name' was not in simplestreams data
On first publish of an image, nothing was publishing the 'name' field.
subsequent runs would get it published (as the read from glance would get
copied in and then updated), but the first would have no 'name' field, unless
it was present in the source.

This puts the 'name' item in as it is registered to glance.

bzr-revno: 258