improve Makefile so that .sjson and .gpg files are created once
previously, 'make gnupg' had to happen before 'make test'.
Now, we mark that make test depends on 'examples-sign'.
and add a phoney rule that handles that.

Instead of running that phoney rule every time, we run some shell
which does essentially the same thing.  The reason we can't do this
properly with dependencies is because of the ':' in our file names.

http://www.mail-archive.com/bug-make@gnu.org/msg03318.html

bzr-revno: 285.1.3