change MirrorReader 'read' to 'read_json'.
This cleans some things up.  read_json is expecting to read json
files, and will call policy on them.  It will call policy with
string content string (not bytes) as it knows what to expect there.

Also, for better or worse, we're now consistently decoding and encoding
with utf-8.  This decode/encode could be moved to util.py to make one
single place where all those happen, but for the moment we don't have
too many places where it is happening.

load_content loads json, and so needs a string so it decodes bytes
if necessary just as a convenience.

bzr-revno: 296