add 'export-tarball' and use that from build-deb.
This pulls in some tool cleanups that I did for curtin.
the export-tarball is nice to be able to cleanly export a nicely
named tarball of trunk or of a tag.  And then build-deb just uses
that.

The move of changelog to changelog.trunk to is allow trunk to have
its own changelog and not conflict with packaged versions.

bzr-revno: 312