attempt trailing / fixup only once and only if necessary.
See bug 1240652 for more details.
The quick open(), close() sequence for url resources seems to trigger
problems in urllib3, at least as used by python-requests and with a proxy
enabled.

It would probably be better to do a HEAD instead of a GET there to do this
check.  Now, we're doing it only once ever per UrlMirrorReader, and only
if the prefix does not end in a /.

The end result is that we do not do it at all on correct input and only
once ever on incorrect input.

LP: #1240652
bzr-revno: 319