tools/make-test-data: get checksums of manifest files
the manifest files are not represented in the MD5SUM and SHA256SUM
files.  So, we pull them and calculate the sum locally.

Also, to improve the cache, we store that we've pulled the
MD5SUM and SHA256SUM files before.  This way we know if we've pulled
the checksum file to search for .manifest before.  rather than assuming
that manifest is never in the cache.

bzr-revno: 358