get_cloud_images_file_info: fix to only return checksums and size
this was returning the content of a manifest, which was then 
getting added.

This makes that method only return desired data independent of
whether it hit from cache or not.

Also, remove the two crawls of the tree, as we have the
md5/sha256 and size all at once now.

bzr-revno: 371.1.6