objectstores must close their readers if they read from them.
This fixes the leaking file handles that was happening when
a 'reader' (in most cases a UrlContentSource) was being handed
to 'insert_item', it would never be closed.

This is possibly confusing because it never was explicitly open()ed.
So the user had a reasonable excuse for not calling close().

The code causing the issue in MAAS was similar, in that the reader
was being passed around, and then read/copied and not 'close()ed'.

LP: #1461181
bzr-revno: 381