tools/make-test-data: include additional info in stream output
This adds the following fields of information to each of the dictionaries
in release data:
  os_title: literal "Ubuntu"
  release_codename: the full code name ('Saucy Salamander', 'Trusty Tahr')
  release_title: numeric portion + " LTS" if this is an lts
                 '13.10', '14.04 LTS"

Each of those fields are added to stream data.

note, 'devel' and 'supported' are only present if it gets data from
the distro_info module.

we add additional fields of information for each of the data in REL2VER.
Then, simplify the hard coded data we have to maintain.

LP: #1475753, #1475754
bzr-revno: 387