remove 'os_title'.
It wasn't clear what we were gaining from this.
The reason is that we may have a product that is 'Ubuntu Server'
and one that is 'Ubuntu Desktop'. Both are 'ubuntu' as the distro
but the 'os_title' doesn't buy us anything here.

Perhaps sometime in the future we'd add 'Ubuntu Server' as a product_title
or something.

bzr-revno: 388