make-test-data: add 'version' back to download data
version was inadvertently dropped from download data in revision 387.
It had stayed in the generated ec2 data.

This adds 'new_ubuntu_product_dict' and uses it in both places
to create the initial product data.  That should ensure that we dont
make download and ec2 different unintentionally in the future.

This also will end up adding 'release_title' and 'release_codename'
to the ec2 data.  And also will add 'supported' for the case
where python-distroinfo is available during the build
(tools/ubuntu_versions.py only gets that data if using distro info
so we are not expected to keep up to date information here).

bzr-revno: 391