make-test-data: support skipping unsupported releases
This adds support for skipping the publication of data for releases
that have reached end of life.  We came to do this work when 
bad data on cloud-images.ubuntu.com (described in bug 1493395) meant
that many referenced urls were not present.

The general idea seems sane though, so we'll add this code.
The behavior can be controlled by environment variable 'SS_SKIP_UNSUPPORTED'
which defaults ot '0'.  Setting to a value other than '0' will mean
that the unsupported releases are simply skipped.

Doing so means we must have distro-info (as it is how we get the
supported flag for a release).

bzr-revno: 396