sync_products: pass a checksumming content source on insert_item
instead of calling insert_item with a bare content source, pass
a checksumming content source that will error if bad data is read.

This adds a way via environment variable to get a non-checking content
source in the event that the user is using a source that wouldnt have all
data.

bzr-revno: 398.1.9