Set product aliases without modifying release data dictionary
In r424 we added support for aliases.  With that code it is possible
that the daily devel products will lack the intended 'devel' alias.

Aliases for a release/arch are set from toolutil.REL2VER[release] and
if the stream is not daily the 'devel' item in the list of aliases is
removed, however the code from that commit accidentally removed 'devel'
from the source dictionary rather than creating a local copy first.
If the devel release is evaluated before the devel daily then the 'devel'
alias will not be present in the dictionary.

This patch will copy the aliases from the release dictionary and
modify the copy for the product being indexed.

bzr-revno: 424.1.1