support sync to glance with image type squashfs for nova-lxd
nova-lxd supports squashfs images, Ubuntu 17.04 does not currently
produce -root.tar.xz (or .gz) images.  Thus we need to support the
squashfs image type.

Also here is a change to start inserting in insert_version rather
than inserting in insert_image directly.  That is so that we have
enough context to only load one of root-tar or squashfs.

LP: #1686086
bzr-revno: 448.2.1