Add support for the Keystone v3 API
simplestreams.openstack is hardcoded for Keystone API v2 client. This change adds
support for Keystone API v3. The version of the API requested is derived from
auth_url and if that fails it falls back to v2.

(This changed borrowed from, with a few changes, from the gnuoy branch that wasn't
merged due to flake errors.)

Closes-Bug: Bug #1624306

bzr-revno: 449.1.1