tox.ini: Pin versions of style check tools used in flake8 env.
As upstream makes changes, new errors can come up causing breakage
in daily c-i that are unrelated to project errors.  The solution here
is to pin the versions of the upstream projects so that style checks
are done on a specific version.

In other projects we have also added a 'flake8-tip' that could be run
but was not run by default.  We can do that at a later date if deemed
necessary.

bzr-revno: 452