Fix ubuntu_versions.py with distro_info > 0.14.
The 'hack_all' nature of 'hack_all' variable name showed itself.
We had been accessing '_rows'.  That is now '_releases' and 
the content there is no longer a dictionary but rather a
DistroRelease object.

The reason we have to do this at all is to get at the datetime
object for 'eol'.  Without that, all we can do is provide a date
and check if a release is supported as of that date.  What we want
is to publish the EOL date (in YYYY-MM-DD format).

This change was present in distro_info at 0.15.  Artful had 0.17,
bionic has 0.18.  Apparently we just didn't notice until now.

bzr-revno: 456.2.1